删除评论功能修改

+++ src/main/java/com/example/service/impl/BookReviewService.java
@@ -1,9 +1,13 @@
 package com.example.service.impl;
 
+import com.example.entity.dto.Account;
+import com.example.entity.vo.request.Book;
 import com.example.mapper.BookReviewMapper;
 import com.example.entity.vo.request.BookReview;
 import com.example.entity.vo.request.CommentDeletionRequest;
 import com.example.entity.vo.request.Page;
+import com.example.service.AccountService;
+import jakarta.annotation.Resource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +16,8 @@ import java.util.List;
 public class BookReviewService {
 
     private final BookReviewMapper bookReviewMapper;
+    @Resource
+    AccountService accountService;
 
     @Autowired
     public BookReviewService(BookReviewMapper bookReviewMapper) {
@@ -53,7 +59,14 @@ public class BookReviewService {
     }
     public boolean deleteCommentsIfUser(CommentDeletionRequest request) {
         // 执行批量删除操作
-        bookReviewMapper.deleteCommentsByIds(request.getIds());
+        for(int i=0;i<request.getIds().size();i++) {
+            BookReview bookReview = bookReviewMapper.selectBookReviewById(request.getIds().get(i));
+            if (bookReview.getUserEmail() == request.getAdminAccount()) {
+                bookReviewMapper.deleteCommentsByIds(request.getIds());
+            } else {
+                return false;
+            }
+        }
         return true;
     }
 }
